[C-2132] unlink is safe on ios but not android (#2852)

+++ packages/mobile/src/services/track-download.ts
@@ -67,6 +67,8 @@ const download = async ({
 
     // On failure attempt to delete the file
     try {
+      const exists = await RNFetchBlob.fs.exists(filePath)
+      if (!exists) return
       await RNFetchBlob.fs.unlink(filePath)
     } catch {}
   }
